Added rm file checker
Former-commit-id: 603c38c79fa932d31c1f2e5e4da3ec5911953193

+++ misc/scripts/upgrade.sh
@@ -26,7 +26,9 @@ function version_gt() {
 }
 
 list=( $(pacstall -L) )
-rm /tmp/pacstall-up-list
+if [ -f /tmp/pacstall-up-list ]; then
+    rm /tmp/pacstall-up-list
+fi
 touch /tmp/pacstall-up-list
 REPO=$(cat /usr/share/pacstall/repo/pacstallrepo.txt)
 fancy_message info "Getting local/remote versions, this may take a while"
